+++ TinyEngine/src/math/Random.cpp
@@ -10,7 +10,7 @@ float Random::radicalInverse(uint32_t bits)
      bits = ((bits & 0x33333333u) << 2u) | ((bits & 0xCCCCCCCCu) >> 2u);
      bits = ((bits & 0x0F0F0F0Fu) << 4u) | ((bits & 0xF0F0F0F0u) >> 4u);
      bits = ((bits & 0x00FF00FFu) << 8u) | ((bits & 0xFF00FF00u) >> 8u);
-     return float(bits) * 2.3283064365386963e-10; // Division by 0x100000000
+     return static_cast<float>(double(bits) * 2.3283064365386963e-10); // Division by 0x100000000
 }
 
 glm::vec2 Random::hammersley(uint32_t i, uint32_t N)
